externalauth 8.x-1.3 -> 8.x-1.4
(Obtained from : https://www.drupal.org/project/externalauth/releases/8.x-1.4) 

Release info
------------

This is a new release for external_auth, which provides some small improvements. 
It does not introduce any major new functionalities or features, and should be safe to update. This will probably be (one of) the last 1.x branch releases for this module.

We are planning to introduce some bigger new features / improvements to the external_auth module though. If you are a module developer that uses this module as a dependency for one of your modules, or you are just interested to help test, review and improve this module, please check the 2.0.x branch issues at [https://www.drupal.org/project/issues/externalauth?status=Open&version=a...] 
Your help and input would be very much appreciated on these issues.

Changelog
---------

### Bug fixes

- [#3213268: Cannot run externalauth_update_8101]

### Test improvements

- [#3143142: Replace assert\* involving an instanceof operator with assertInstanceOf()/assertNotInstanceOf()], Assigned to: deulenko")
- [#3136157: \\Drupal\\externalauth\\Tests\\AuthmapTest should be in namespace \\Drupal\\Tests\\externalauth\\Kernel]")

### Documentation improvements

- [#3134256: Document FALSE as a return value for AuthmapInterface::get()]")
- [#3217524: Use 'User authentication' package instead of 'Authentication']